Move signature values generation to template.
