+++ src/ddo/ServiceAgreementTemplate.ts
@@ -1,7 +1,7 @@
 export interface ServiceAgreementTemplateParameter {
     name: string
     type: string
-    value: string
+    value: string | number
 }
 
 export interface ServiceAgreementTemplateEvent {
