+++ src/ocean/OceanAgreements.ts
@@ -3,7 +3,10 @@ import BrizoProvider from "../brizo/BrizoProvider"
 import { generateId } from "../utils/GeneratorHelpers"
 import Account from "./Account"
 import DID from "./DID"
+import { DDO } from "../ddo/DDO"
 import ServiceAgreement from "./ServiceAgreements/ServiceAgreement"
+import { Keeper } from "../keeper/Keeper"
+import { zeroX, didPrefixed } from "../utils"
 
 import OceanAgreementsConditions from "./OceanAgreementsConditions"
 
@@ -54,44 +57,20 @@ export default class OceanAgreements {
         serviceDefinitionId: string,
         consumer: Account,
     ): Promise<AgreementPreparionResult> {
+        const keeper = await Keeper.getInstance()
+
         const d: DID = DID.parse(did as string)
         const ddo = await AquariusProvider.getAquarius().retrieveDDO(d)
         const agreementId: string = generateId()
 
-        const valuesMap = await this.getValuesMapForPrepare(ddo, agreementId, consumer.getId())
+        const templateName = ddo.findServiceByType("Access").serviceAgreementTemplate.contractName
+        const valuesMap = await keeper
+            .getTemplateByName(templateName)
+            .getServiceAgreementTemplateValuesMap(ddo, agreementId, consumer.getId())
 
         const signature = await ServiceAgreement.signServiceAgreement(ddo, serviceDefinitionId, agreementId, valuesMap, consumer)
 
         return {agreementId, signature}
-    }
-
-    // TODO: this map depends on the template, this generation should be reponsability of the template
-    private async getValuesMapForPrepare(ddo: DDO, agreementId: string, consumer: string): Promise<{[value: string]: string}> {
-        const keeper = await Keeper.getInstance()
-        const ddoOwner = ddo.proof && ddo.proof.creator
-        const amount = ddo.findServiceByType("Metadata").metadata.base.price
-
-        let lockCondition
-        let releaseCondition
-
-        try {
-            lockCondition = await keeper.conditions.lockRewardCondition.hashValues(ddoOwner, amount)
-        } catch(e) { }
-
-        try {
-            releaseCondition = await keeper.conditions.accessSecretStoreCondition.hashValues(ddo.shortId(), consumer)
-        } catch(e) { }
-
-        return {
-            rewardAddress: ddoOwner,
-            amount: amount.toString(),
-            documentId: ddo.shortId(),
-            grantee: consumer,
-            receiver: consumer,
-            sender: ddoOwner,
-            lockCondition,
-            releaseCondition,
-        }
     }
 
     /**
@@ -147,16 +126,19 @@ export default class OceanAgreements {
         const d: DID = DID.parse(did)
         const ddo = await AquariusProvider.getAquarius().retrieveDDO(d)
 
-        const serviceAgreement: ServiceAgreement = await ServiceAgreement
-            .executeServiceAgreement(
-                d,
-                ddo,
-                agreementId,
-                serviceDefinitionId,
-                signature,
-                consumer,
-                publisher)
+        // TODO: Use Keeper 0.7+
 
-        return serviceAgreement
+        // const serviceAgreement: ServiceAgreement = await ServiceAgreement
+        //     .executeServiceAgreement(
+        //         d,
+        //         ddo,
+        //         agreementId,
+        //         serviceDefinitionId,
+        //         signature,
+        //         consumer,
+        //         publisher)
+
+        // return serviceAgreement
+        return undefined
     }
 }
